// src/features/Authentication/model/enums.ts
export enum AuthStages {
    NOT_AUTHENTICATED = 'NOT_AUTHENTICATED',
    AUTHENTICATING = 'AUTHENTICATING',
}

export enum AuthTransitionTypes {
    // Inner
    NOT_AUTHENTICATED__AUTHENTICATING = 'NOT_AUTHENTICATED__AUTHENTICATING',
    AUTHENTICATING__NOT_AUTHENTICATED = 'AUTHENTICATING__NOT_AUTHENTICATED',
    // To product list
    AUTHENTICATING__LOADING_PRODUCTS = 'AUTHENTICATING__LOADING_PRODUCTS',
}